tests/wpt/web-platform-tests/font-access/permissions-policy/local-fonts-supported-by-permissions-policy.html

<!DOCTYPE html>
<title>Test that local fonts is advertised in the feature list</title>
<link rel="help" href="https://w3c.github.io/webappsec-permissions-policy/#dom-permissions-policy-features">
<link rel="help" href="https://wicg.github.io/local-fonts/#permissions-policy">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
  test(() => {
      assert_in_array('local-fonts', document.featurePolicy.features());
  }, 'document.featurePolicy.features should advertise local-fonts.');
</script>
